python_scripts/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LangGraph & LangChain
langgraph==0.0.62
langchain==0.1.0
langchain-openai==0.0.2

# AI & Embeddings
openai==1.6.1

# Vector Database
chromadb==0.4.18

# Data Processing
pandas==2.1.4
numpy==1.24.3

# Web Requests (for Naver News API)
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Text Processing
tiktoken==0.5.2

# Environment
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging
loguru==0.7.2